backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
httpx==0.27.2
aiogram==3.12.0
openai==1.54.3
ccxt==4.3.96
websockets==13.0
polars==1.6.0
duckdb==1.1.2
xgboost==2.1.1
scikit-learn==1.5.2
# PyTorch (opsiyonel). Apple Silicon/arm64 için uygun sürümü kullanalım.
torch==2.8.0
redis==5.0.8
sqlalchemy==2.0.35
prometheus-client==0.21.0
python-dotenv==1.0.1
feedparser==6.0.11
rich==13.9.2
typer==0.12.5
PyYAML==6.0.2
zstandard==0.23.0
boto3==1.35.23
APScheduler==3.10.4
web3==6.20.1
networkx==3.3
requests==2.32.3

# ML
scikit-learn==1.5.2
joblib==1.4.2

# Telegram
telethon>=1.34.0
aiohttp>=3.9.0

# test
pytest==8.3.3
pytest-asyncio==0.24.0
httpx>=0.27.0
pytest-benchmark>=4.0.0

# redis (distributed rate limit)
redis>=5.0.0

# TimescaleDB / PostgreSQL async
asyncpg==0.30.0
psycopg2-binary==2.9.9

# SSE Client (for health checks)
sseclient-py==1.8.0

# s3 archiver
boto3>=1.34.0

pandas==2.2.3
